--- a/pkg/traceql/ast_execute.go
+++ b/pkg/traceql/ast_execute.go
@@ -110,7 +110,7 @@ func (o SpansetOperation) evaluate(input []*Spanset) (output []*Spanset, err err
 			fallthrough
 		case OpSpansetDescendant:
 			relFn = func(l, r []Span) []Span {
-				return l[0].DescendantOf(r, l, falseForAll, true, o.matchingSpansBuffer) // jpe - l[0] is bad
+				return l[0].DescendantOf(l, r, falseForAll, true, o.matchingSpansBuffer) // jpe - l[0] is bad
 			}
 
 		case OpSpansetNotAncestor:
@@ -126,7 +126,7 @@ func (o SpansetOperation) evaluate(input []*Spanset) (output []*Spanset, err err
 			fallthrough
 		case OpSpansetChild:
 			relFn = func(l, r []Span) []Span {
-				return l[0].ChildOf(r, l, falseForAll, true, o.matchingSpansBuffer)
+				return l[0].ChildOf(l, r, falseForAll, true, o.matchingSpansBuffer)
 			}
 
 		case OpSpansetNotParent: