--- a/pkg/traceql/ast_execute.go
+++ b/pkg/traceql/ast_execute.go
@@ -151,7 +151,7 @@ func (o SpansetOperation) evaluate(input []*Spanset) (output []*Spanset, err err
 
 		// if relFn was set up above we are doing a relationship operation.
 		if relFn != nil {
-			o.matchingSpansBuffer, err = o.joinSpansets(lhs, rhs, falseForAll, relFn)
+			o.matchingSpansBuffer, err = o.joinSpansets(lhs, rhs, relFn)
 			if err != nil {
 				return nil, err
 			}
@@ -172,7 +172,7 @@ func (o SpansetOperation) evaluate(input []*Spanset) (output []*Spanset, err err
 // joinSpansets compares all pairwise combinations of the inputs and returns the right-hand side
 // where the eval callback returns true.  For now the behavior is only defined when there is exactly one
 // spanset on both sides and will return an error if multiple spansets are present.
-func (o *SpansetOperation) joinSpansets(lhs, rhs []*Spanset, falseForAll bool, eval func(l, r []Span) []Span) ([]Span, error) {
+func (o *SpansetOperation) joinSpansets(lhs, rhs []*Spanset, eval func(l, r []Span) []Span) ([]Span, error) {
 	if len(lhs) < 1 || len(rhs) < 1 {
 		return nil, nil
 	}