--- a/pkg/traceqlmetrics/mocks.go
+++ b/pkg/traceqlmetrics/mocks.go
@@ -45,13 +45,19 @@ func (m *mockSpan) WithErr() *mockSpan {
 	return m
 }
 
-func (m *mockSpan) Attributes() map[traceql.Attribute]traceql.Static                 { return m.attrs }
-func (m *mockSpan) ID() []byte                                                       { return nil }
-func (m *mockSpan) StartTimeUnixNanos() uint64                                       { return m.start }
-func (m *mockSpan) DurationNanos() uint64                                            { return m.duration }
-func (m *mockSpan) DescendantOf([]traceql.Span, []traceql.Span, bool) []traceql.Span { return nil }
-func (m *mockSpan) SiblingOf([]traceql.Span, []traceql.Span, bool) []traceql.Span    { return nil }
-func (m *mockSpan) ChildOf([]traceql.Span, []traceql.Span, bool) []traceql.Span      { return nil }
+func (m *mockSpan) Attributes() map[traceql.Attribute]traceql.Static { return m.attrs }
+func (m *mockSpan) ID() []byte                                       { return nil }
+func (m *mockSpan) StartTimeUnixNanos() uint64                       { return m.start }
+func (m *mockSpan) DurationNanos() uint64                            { return m.duration }
+func (m *mockSpan) DescendantOf([]traceql.Span, []traceql.Span, bool, bool, []traceql.Span) []traceql.Span {
+	return nil
+}
+func (m *mockSpan) SiblingOf([]traceql.Span, []traceql.Span, bool, bool, []traceql.Span) []traceql.Span {
+	return nil
+}
+func (m *mockSpan) ChildOf([]traceql.Span, []traceql.Span, bool, bool, []traceql.Span) []traceql.Span {
+	return nil
+}
 
 type mockFetcher struct {
 	filter   traceql.SecondPassFn