--- a/pkg/traceqlmetrics/mocks.go
+++ b/pkg/traceqlmetrics/mocks.go
@@ -52,7 +52,7 @@ func (m *mockSpan) DurationNanos() uint64                            { return m.
 func (m *mockSpan) DescendantOf([]traceql.Span, []traceql.Span, bool, bool, []traceql.Span) []traceql.Span {
 	return nil
 }
-func (m *mockSpan) SiblingOf([]traceql.Span, []traceql.Span, bool, bool, []traceql.Span) []traceql.Span {
+func (m *mockSpan) SiblingOf([]traceql.Span, []traceql.Span, bool, []traceql.Span) []traceql.Span {
 	return nil
 }
 func (m *mockSpan) ChildOf([]traceql.Span, []traceql.Span, bool, bool, []traceql.Span) []traceql.Span {