… pkg/traceql/ast_test.go
… tempodb/encoding/vparquet2/combiner_test.go
… tempodb/encoding/vparquet2/nested_set_model.go
… tempodb/encoding/vparquet2/nested_set_model_test.go
… tempodb/encoding/vparquet2/schema_test.go
… tempodb/encoding/vparquet3/block_traceql.go
… tempodb/encoding/vparquet3/block_traceql_test.go
… tempodb/encoding/vparquet3/combiner_test.go
… tempodb/encoding/vparquet3/nested_set_model.go
… tempodb/encoding/vparquet3/nested_set_model_test.go
… tempodb/encoding/vparquet3/schema_test.go
… tempodb/tempodb_search_test.go