**What this PR does**:
The PR improves performance by using a standard sort/search procedure but that requires having both the LHS and RHS of the operator at the fetch layer (where it is aware of nested set implementation details). The cost is a very ungainly change to the span interface:

```
	SiblingOf([]Span, []Span, bool, bool, []Span) []Span
	DescendantOf([]Span, []Span, bool, bool, []Span) []Span
	ChildOf([]Span, []Span, bool, bool, []Span) []Span
```

These methods no longer operate on the span itself which is strange. However, there is no other clean places to place it that can be easily accessed deep in the engine where the evaluations occur.

Other fixes:
- Fixed an issue where `{} !~ {}` would not find the root span. The way the fix is implemented multiple root spans will be considered siblings using a query like `{} ~ {}`
- Fixed the bridge iterator sort to use the correct definition level

Benchmarks to follow.

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`